chore(evals): move eval scenario declarations into per-language files
Complete the self-registration pattern the registerScenarios mechanism
exists for: each language's scenario now lives in its own
instr_<language>.go file (Python, Java, Ruby, PHP, Scala, Next.js) with
its ID constant, service-name constant, scenario func, and its own
init() call, matching the pattern instr_dotnet.go and instr_browser.go
already followed for their declarations. Those two files now register
their own scenarios too, instead of relying on the central init in
instr_sdks.go.

instr_sdks.go stays as the shared-helpers file for the SDK scenarios
(heavyScenarioTimeout and the sdkHTTPScenario builder), because .NET
reuses both and the package already keeps shared declarations for the
base set in scenarios.go — which "never gets edited again" by design.

This is a pure move: scenario IDs, prompts, rule files, fixtures,
timeouts, and assertions are unchanged (verified by diffing the
serialized scenario data of All() between main and this branch — byte
identical). Registration order shifts from one central call to lexical
file-init order, which register.go documents as the mechanism and which
the order-insensitive enumeration tests (ElementsMatch) permit.

On the issue's second reading (skill content): the shared
testing/verification guidance in
skills/otel-instrumentation/rules/validation.md is language-neutral
(Collector/backend checks), so there is nothing to move into sdks/*.md;
that reading is already satisfied.

Fixes #131
